Add v*-local from upstream
This adds the schema files of the "local" variety to the repo, from
the upstream https://github.com/instrumenta/kubernetes-json-schema.

We only want these files, and the download / checkout of the whole
thing is enormous.

Method:

    git checkout upstream/master
    git checkout --orphan local-only
    git reset
    git add v*-local
    git commit
    # and we're here

(NB: probably not the most effective way to achieve that. Answers on a
postcard ..)